Move testing rudder-parsers to rudder-core due to problems with dependencies
